Create a small regression test for retrieving incorrect records
There is an issue in the constant table saver where if no binds are set,
it checks if the sql query generated is equal to the `all` query.

This `all` query uses the current_scope if its set, otherwise it uses the
defaulted_scope. In this case, current_scope is set, and the `all` query
returns the current scope, which is the same as the sql query.

Since they are the same, super(relation.to_sql) gets called, but
this will retrieve all the relations not just the ones selected by the sql statement.